+++ common.py
@@ -1,13 +1,13 @@
 import logging
 import os
+import smtplib
 import sys
+from email.message import EmailMessage
 from logging.handlers import RotatingFileHandler
 
 from dotenv import load_dotenv
 
 __all__ = ["logger", "API_KEY", "GET_URL", "SERVERS_URL", "MAX_RETRIES", "RETRY_BACKOFF_FACTOR", "ROTATE", "POST_BACKUP_SCRIPT"]
-
-from alert import EmailAlert
 
 # Environment .env
 load_dotenv()
@@ -40,10 +40,59 @@ API_KEY = os.getenv("API_KEY") or ""
 GET_URL = os.getenv("GET_URL") or ""
 SERVERS_URL = os.getenv("SERVERS_URL") or ""
 MAX_RETRIES = int(os.getenv("MAX_RETRIES") or "") or 5
-RETRY_BACKOFF_FACTOR = int(os.getenv("RETRY_BACKOFF_FACTOR") or "") or 1
+RETRY_BACKOFF_FACTOR = int(os.getenv("RETRY_BACKOFF_FACTOR") or "0") or 1
 SEND_EMAILS = (str(os.getenv("SEND_EMAILS") or "").lower() or "true") == "true"
 ROTATE = (str(os.getenv("ROTATE") or "").lower() or "false") == "true"
 POST_BACKUP_SCRIPT = os.getenv("POST_BACKUP_SCRIPT")   # Optional
+
+
+
+class EmailAlert:
+    def __init__(self, from_email, from_password, smtp_server, smtp_port):
+        self.from_email = from_email
+        self.from_password = from_password
+        self.smtp_server = smtp_server
+        self.smtp_port = int(smtp_port)
+
+    def anonymize_email(self, email):
+        local_part, domain = email.split("@")
+        return f"{'*' * len(local_part)}@{domain}"
+
+    def send(self, subject, body, to_email):
+        if not all(
+                isinstance(i, str)
+                for i in [subject, body, to_email, self.from_email, self.from_password]
+        ):
+            raise ValueError("All parameters must be of type str")
+
+        logger.info(f"Preparing to send email to {self.anonymize_email(to_email)}")
+
+        # Create the email message
+        msg = EmailMessage()
+        msg.set_content(body)
+        msg["Subject"] = subject
+        msg["From"] = self.from_email
+        msg["To"] = to_email
+
+        try:
+            # Connect to the mail server using TLS
+            with smtplib.SMTP(self.smtp_server, self.smtp_port) as server:
+                # Enable TLS encryption
+                server.starttls()
+                # Log in to your email account
+                server.login(self.from_email, self.from_password)
+
+                # Send the email
+                server.send_message(msg)
+                logger.info(f"Email sent to {self.anonymize_email(to_email)}")
+        except smtplib.SMTPException as e:
+            logger.error(
+                f"SMTP error occurred when sending email to {self.anonymize_email(to_email)}: {e}"
+            )
+        except Exception as e:
+            logger.error(
+                f"Unexpected error occurred when sending email to {self.anonymize_email(to_email)}: {e}"
+            )
 
 
 # Instantiate EmailAlert object
