+++ requirements.txt
@@ -1,2 +1,3 @@
 requests
 python-dotenv
+urllib3
